# openapi.yaml
openapi: 3.0.3

info:
  title: Shipment Management API
  description: HTTP API for managing shipments and their stops.
  version: 1.0.0

servers:
  - url: http://localhost:5050
    description: Shipment service

tags:
  - name: Shipments
    description: Shipment management operations
  - name: Stops
    description: Shipment stop lifecycle operations

paths:
  /shipments:
    post:
      tags:
        - Shipments
      summary: Create a shipment
      operationId: createShipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShipmentRequest"
            example:
              stops:
                - id: "550e8400-e29b-41d4-a716-446655440000"
                  sequence: 1
                  type: "PICKUP"
                - id: "550e8400-e29b-41d4-a716-446655440001"
                  sequence: 2
                  type: "DELIVERY"
      responses:
        "201":
          description: Shipment created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Shipment violates a domain rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /shipments/{shipmentId}:
    get:
      tags:
        - Shipments
      summary: Get shipment by ID
      operationId: getShipmentById
      parameters:
        - name: shipmentId
          in: path
          required: true
          description: Unique identifier of the shipment
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440010"
      responses:
        "200":
          description: Shipment retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shipment"
        "400":
          description: Invalid shipment ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Shipment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /shipments/{shipmentId}/stops/{stopId}/arrive:
    post:
      tags:
        - Stops
      summary: Mark a stop as arrived
      operationId: arriveAtStop
      parameters:
        - name: shipmentId
          in: path
          required: true
          description: Unique identifier of the shipment
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440010"
        - name: stopId
          in: path
          required: true
          description: Unique identifier of the stop
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Stop marked as arrived successfully
        "400":
          description: Invalid request parameter format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Shipment or stop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Domain rule violation (e.g. previous stops not departed, stop already arrived)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /shipments/{shipmentId}/stops/{stopId}/pickup:
    post:
      tags:
        - Stops
      summary: Pick up freight at a stop
      operationId: pickupAtStop
      parameters:
        - name: shipmentId
          in: path
          required: true
          description: Unique identifier of the shipment
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440010"
        - name: stopId
          in: path
          required: true
          description: Unique identifier of the stop
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: Stop marked as picked up (departed) successfully
        "400":
          description: Invalid request parameter format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Shipment or stop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Domain rule violation (e.g. stop is not a pickup stop, stop not arrived, shipment already completed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

  /shipments/{shipmentId}/stops/{stopId}/deliver:
    post:
      tags:
        - Stops
      summary: Deliver freight at a stop
      operationId: deliverAtStop
      parameters:
        - name: shipmentId
          in: path
          required: true
          description: Unique identifier of the shipment
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440010"
        - name: stopId
          in: path
          required: true
          description: Unique identifier of the stop
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440001"
      responses:
        "200":
          description: Stop marked as delivered (departed) successfully
        "400":
          description: Invalid request parameter format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Shipment or stop not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Domain rule violation (e.g. stop is not a delivery stop, stop not arrived, shipment already completed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

components:
  schemas:
    CreateShipmentRequest:
      type: object
      required:
        - stops
      properties:
        stops:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateShipmentStop"

    CreateShipmentStop:
      type: object
      required:
        - id
        - sequence
        - type
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        sequence:
          type: integer
          minimum: 1
          example: 1
        type:
          type: string
          enum:
            - pickup
            - delivery
          example: pickup

    Shipment:
      type: object
      required:
        - id
        - status
        - stops
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440010"
        status:
          type: string
          enum:
            - IN_TRANSIT
            - COMPLETED
          example: IN_TRANSIT
        stops:
          type: array
          items:
            $ref: "#/components/schemas/Stop"

    Stop:
      type: object
      required:
        - id
        - sequence
        - type
        - status
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        sequence:
          type: integer
          example: 1
        type:
          type: string
          enum:
            - pickup
            - delivery
          example: pickup
        status:
          type: string
          enum:
            - in-transit
            - arrived
            - departed
          example: in-transit

    ProblemDetails:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: "2026-09-11T13:25:00.000Z"
        path:
          type: string
          example: "/shipments"
        method:
          type: string
          example: "POST"
        error:
          type: string
          example: "BadRequestException"
        message:
          oneOf:
            - type: string
              example: "Shipment not found"
            - type: array
              items:
                type: string
              example: ["stops.0.id must be a UUID"]